backend/requirements.txt
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
elastic-transport==8.4.0
elasticsearch==8.7.0
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
importlib-metadata==6.3.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytest-flask==1.2.0
python-dotenv==1.0.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
